sql/ai-embedding-tx-lab/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sentence-transformers
streamlit
requests